requirements.txt
langchain-community
python-dotenv
streamlit
langchain-core
huggingface-hub 